- significant code optimization and memory reduction - saved files are now smaller - collada exporter is now a little bit faster - selection of huge objects is now faster (more optimization is still planned) - loading of projects is now faster - removed lots of unnecessary code

+++ PS4k/src/com/vitco/export/ColladaFile.java
@@ -135,7 +135,7 @@ public class ColladaFile {
     }
 
     // add planes (public wrapper)
-    public void addPlane(float[] pos, int orientation,
+    public void addPlane(int[] pos, int orientation,
                          Color color, Integer textureId, Integer rotation, Boolean flip) {
         // create the plane material
         PlaneMaterial material = new PlaneMaterial(color, textureId, rotation, flip, orientation);
